sidebar - separated scope and namespace tree nodes
