--- components/side-bar/component-tree/component-nodes/index.ts
@@ -1,1 +1,0 @@
-export * from './component-nodes';
