+++ components/side-bar/component-tree/index.ts
@@ -1,6 +1,7 @@
 export * from './component-tree';
 export * from './recursive-tree';
 export { ComponentView } from './component-view';
-export { ScopeView, NamespaceView } from './component-nodes';
+export { ScopeTreeNode } from './scope-tree-node';
+export { NamespaceTreeNode } from './namespace-tree-node';
 export { ScopePayload } from './payload-type';
 export type { PayloadType } from './payload-type';
