+++ components/side-bar/component-tree/namespace-tree-node/index.ts
@@ -1,0 +1,1 @@
+export { NamespaceTreeNode } from './namespace-tree-node';
