+++ components/side-bar/component-tree/namespace-tree-node/namespace-tree-node.module.scss
