+++ components/side-bar/component-tree/namespace-tree-node/namespace-tree-node.tsx
@@ -8,32 +8,9 @@ import { indentClass, indentStyle } from '../indent';
 import { PayloadType } from '../payload-type';
 import { TreeLayer, TreeNodeProps } from '../recursive-tree';
 import { getName } from '../utils/get-name';
-import styles from './component-nodes.module.scss';
+import styles from './namespace-tree-node.module.scss';
 
-export function ScopeView({ node, depth }: TreeNodeProps<PayloadType>) {
-  const [collapsed, collapse] = useState(false);
-  const displayName = getName(node.id.replace(/\/$/, ''));
-
-  return (
-    <div>
-      {node.id && (
-        <div className={classNames(indentClass, clickable, styles.namespace)} onClick={() => collapse(!collapsed)}>
-          <div className={styles.left}>
-            <Icon className={classNames(styles.arrow, collapsed && styles.collapsed)} of="fat-arrow-down" />
-            <Icon className={styles.arrow} of="collection-full" />
-            <span className={styles.name}>{displayName}</span>
-          </div>
-        </div>
-      )}
-      <AnimateHeight height={collapsed ? 0 : 'auto'}>
-        <div style={indentStyle(depth + 1)} className={classNames(styles.componentTree)}>
-          {node.children && <TreeLayer childNodes={node.children} depth={depth + 1} />}
-        </div>
-      </AnimateHeight>
-    </div>
-  );
-}
-export function NamespaceView({ node, depth }: TreeNodeProps<PayloadType>) {
+export function NamespaceTreeNode({ node, depth }: TreeNodeProps<PayloadType>) {
   const [collapsed, collapse] = useState(false);
 
   const displayName = getName(node.id.replace(/\/$/, ''));
