+++ components/side-bar/component-tree/scope-tree-node/index.ts
@@ -1,0 +1,1 @@
+export { ScopeTreeNode } from './scope-tree-node';
