+++ components/side-bar/index.ts
@@ -1,4 +1,11 @@
 export { ComponentTree } from './component-tree';
 export { Collapser } from './sidebar-collapser';
 export { StatusTypes } from './component-tree/recursive-tree';
-export { ComponentView, ScopeView, NamespaceView, PayloadType, TreeNodeProps, ScopePayload } from './component-tree';
+export {
+  ComponentView,
+  ScopeTreeNode,
+  NamespaceTreeNode,
+  PayloadType,
+  TreeNodeProps,
+  ScopePayload,
+} from './component-tree';
